old/legacy-root-restored/packages/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
motor>=3.6.0
pydantic-settings>=2.9.0
PyJWT>=2.10.1
email-validator>=2.2.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.2.0
scipy>=1.10.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
# Component 4's BiLSTM-CRF model is optional and only works on Python versions
# supported by TensorFlow.
tensorflow>=2.16.0; python_version < "3.13"
openpyxl>=3.0.0
requests>=2.28.0
tqdm>=4.65.0
